+++ matrix-meetings-bot/src/service/MeetingService.ts
@@ -63,8 +63,8 @@ import { StateEventName } from '../model/StateEventName';
 import { WidgetType } from '../model/WidgetType';
 import { IMeetingChanges, meetingChangesHelper } from '../util/IMeetingChanges';
 import { templateHelper } from '../util/TemplateHelper';
-import { getForceDeletionTime } from '../util/calendarUtils';
 import { extractOxRrule } from '../util/extractOxRrule';
+import { getForceDeletionTime } from '../util/getForceDeletionTime';
 import { migrateMeetingTime } from '../util/migrateMeetingTime';
 import { RoomMessageService } from './RoomMessageService';
 import { WidgetLayoutService } from './WidgetLayoutService';
