+++ matrix-meetings-bot/src/util/TemplateHelper.ts
@@ -14,12 +14,14 @@
  * limitations under the License.
  */
 
-import { parseICalDate } from '@nordeck/matrix-meetings-calendar';
+import {
+  getSingleOrRecurringEntry,
+  parseICalDate,
+} from '@nordeck/matrix-meetings-calendar';
 import i18next from 'i18next';
 import { fullNumericDateFormat } from '../dateFormat';
 import { CalendarEntryDto } from '../dto/CalendarEntryDto';
 import { IUserContext } from '../model/IUserContext';
-import { getSingleOrRecurringEntry } from './calendarUtils';
 import { formatRRuleText } from './format';
 
 export class TemplateHelper {
