+++ packages/calendar/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@nordeck/matrix-meetings-calendar",
-  "description": "A calendar utils library used by widget and bot",
+  "description": "A calendar utils library",
   "license": "Apache-2.0",
   "private": true,
   "version": "0.0.1",
