+++ packages/calendar/src/utils/calendarUtils/helpers.ts
@@ -75,3 +75,21 @@ export function createTimeFilter(fromDate: string, toDate?: string) {
     );
   };
 }
+
+/**
+ * Extracts first single or recurring calendar entry.
+ * @param calendar meeting calendar
+ */
+export function getSingleOrRecurringEntry(
+  calendar: CalendarEntry[],
+): CalendarEntry {
+  const entry = calendar.find((e) => !isRRuleOverrideEntry(e));
+
+  if (entry === undefined) {
+    throw new Error(
+      'Unexpected input: calendar must have single or recurring entry',
+    );
+  }
+
+  return entry;
+}
